Refactor utils into multiple files
